requirements.txt
asgiref==3.9.1
click==8.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
PyJWT==2.10.1
qrcode==8.2
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.36.0
whitenoise==6.11.0
